feat: add convert for integers array

+++ src/main/java/com/daguinci/menumaker_api/IngredientRepository.java
@@ -7,6 +7,6 @@ import com.daguinci.menumaker_api.Ingredient;;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface IngredientRepository extends CrudRepository<Ingredient, Integer> {
+public interface IngredientRepository extends CrudRepository<Ingredient, Long> {
     boolean existsByName(String name);
 }
